instagram/backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
PyMySQL>=1.1,<2.0
cryptography>=42.0
python-decouple>=3.8
celery>=5.3,<6.0
redis>=5.0,<6.0
pytest>=8.0
pytest-django>=4.8
requests>=2.32
